Add swager, refactor tasks service

+++ src/tasks/tasks.service.ts
@@ -1,32 +1,44 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException, UseGuards } from '@nestjs/common';
 import { Prisma, PrismaClient, Task  } from '@prisma/client';
 import { PrismaService } from 'src/prisma.service';
+import { createTask } from './dto/create-task-dto';
+import { AuthGuard } from 'src/auth/auth.guard';
 
 @Injectable()
 export  class TasksService {
     constructor(private prisma: PrismaService) {}
+    
     async getAll(){
         return await this.prisma.task.findMany()
     }
 
     async getById(id : number) {
-       return await this.prisma.task.findUnique({where: {id: id}})
+       const taskFound = await this.prisma.task.findUnique({where: {id: id}})
+        if(!taskFound) { throw new NotFoundException('Tarea no encontrada') }
+        return taskFound
     }
 
-    create(task: any){
-        return this.prisma.task.create({data: task})
+    async create(task: createTask){
+        return await this.prisma.task.create({data: task})
     }
 
-    delete(id: number){
+    async delete(id: number){
+        const taskFound = await this.prisma.task.findUnique({where: {id: id}})
+        if(!taskFound) { throw new NotFoundException('Tarea no encontrada') }
         return this.prisma.task.delete({where: {id: id}})
     }
 
-    update(id: number, task: any) {
+    async update(id: number, task: Task) {
+        const taskFound = await this.prisma.task.findUnique({where: {id: id}})
+        if(!taskFound) { throw new NotFoundException('Tarea no encontrada') }
         return this.prisma.task.update({where: {id: id}, data: task})
     }
 
-    patchStatus(id: number, status: string) {
-        return this.prisma.task.update({where: {id: id}, data: {status: status}})
+    async changeStatus(id: number, data: {status: string}) {
+        return this.prisma.task.update({where: {id: id}, data: data})
     }
 
+    async changePriority(id: number, data: {priority: string}) {
+        return this.prisma.task.update({where: {id: id}, data: data})
+    }
 }
